backend/requirements.txt
# Core
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.9
asyncpg==0.28.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0.post2

# Data Validation
pydantic==2.3.0
pydantic-settings==2.0.3

# Async
anyio==3.7.1
httpx==0.24.1

# Utils
python-dateutil==2.8.2
python-slugify==8.0.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1

# Documentation
fastapi-pagination==0.12.8

# Email
aiosmtplib==2.0.1

# File handling
aiofiles==23.2.1

# Logging
structlog==23.1.0

# CORS
starlette==0.27.0

# Development
black==23.7.0
isort==5.12.0
mypy==1.5.1
pylint==2.17.5
